type conversion to long in comparison removed
In the previous version we could have possibly an unexpected infinite loop: if `long` can represent larger values than `double`, `evo->gen` might never exceed `startiter + iterations`. In the updated version we will never do more iterations that either `long` or `double` can represent. Now, the absolute number of iterations remain always defined as `double`, the added iterations are `long`.

+++ src/cmaes.c
@@ -763,10 +763,10 @@ cmaes_Optimize( cmaes_t *evo, double(*pFun)(double const *, int dim), long itera
     double *const*pop; /* sampled population */
     const char *stop;
     int i;
-    long startiter = (long)evo->gen; 
+    double startiter = evo->gen; 
     
     while(!(stop=cmaes_TestForTermination(evo)) && 
-        ((long)evo->gen < startiter + iterations || !iterations))
+        (evo->gen < startiter + iterations || !iterations))
     { 
         /* Generate population of new candidate solutions */
         pop = cmaes_SamplePopulation(evo); /* do not change content of pop */
